Do not backup symlinks and add README
This prevents actual files from being overwritten the second time that
dotfiles are installed.

+++ install_dotfiles.sh
@@ -13,8 +13,10 @@ cd $dir
 echo "...done"
 
 for file in $files; do
-  echo "Moving existing dotfiles from ~ to $olddir"
-  mv ~/.$file $olddir
+  if [ -f ~/.$file ] && [ ! -h ~/.$file ]; then
+    echo "Moving ~/.$file from ~ to $olddir"
+    mv ~/.$file $olddir
+  fi
   echo "Creating symlink to $file in home directory."
   ln -s $dir/$file ~/.$file
 done
